requirements.txt
Flask
Flask-Login
Flask-Cors
Flask-SQLAlchemy
#psycopg2
psycopg2cffi
tornado
requests
